requirements.txt
pycryptodome>=3.19.1
pytest>=7.4.0